LPD-88886 Move logic to a new class
