+++ modules/util/source-formatter/source-formatter.properties
@@ -44,6 +44,8 @@
 
     source.check.JavaMetaAnnotationsCheck.checkMissingName=true
 
+    source.check.JavaMetaOCDSecretFieldCheck.enabled=true
+
     source.check.JavaModuleInheritableVariableAccessModifierCheck.enabled=false
 
     source.check.JavaNewProblemInstantiationParametersCheck.enabled=true
