+++ source-formatter.properties
@@ -1276,7 +1276,7 @@
 
     source.check.JavaMetaAnnotationsCheck.checkMissingName=true
 
-    source.check.JavaMetaAnnotationsCheck.checkMissingPasswordType=true
+    source.check.JavaMetaOCDSecretFieldCheck.enabled=true
 
     source.check.JavaModuleClassGetResourceCallCheck.enabled=true
